Update and simplify POM
Specify desired Java version in a more simple manner than before.
Do the same with JUnit 5.

In earlier versions of IntelliJ the Java version had to be specified in
the previou version or it seemed so. The maven-surefire-plugin's earlier
versions wasn't able to detect and run JUnit 5 tests from the
command-line, but this works with the included version.